requirements.txt
accelerate==0.29.2
auto_mix_prep==0.2.0
colorama==0.4.6
datasets==2.18.0
pandas==2.2.2
peft==0.10.0
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.2.2
transformers==4.39.3
trl==0.8.3
wandb==0.16.6
